+++ lib/views/article_view.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_ecoapp/models/article.dart';
 import 'package:flutter_ecoapp/utils/currency_util.dart';
 import 'package:flutter_ecoapp/views/style/colors.dart';
-import 'package:flutter_ecoapp/views/widgets/articles/articleview/description_section.dart';
-import 'package:flutter_ecoapp/views/widgets/articles/articleview/eco_detail_section.dart';
-import 'package:flutter_ecoapp/views/widgets/articles/articleview/photo_section.dart';
-import 'package:flutter_ecoapp/views/widgets/articles/articleview/question_section.dart';
-import 'package:flutter_ecoapp/views/widgets/articles/articleview/store_description_section.dart';
+import 'package:flutter_ecoapp/views/widgets/articles/articleview/article_description_section.dart';
+import 'package:flutter_ecoapp/views/widgets/articles/articleview/article_eco_detail_section.dart';
+import 'package:flutter_ecoapp/views/widgets/articles/articleview/article_photo_section.dart';
+import 'package:flutter_ecoapp/views/widgets/articles/articleview/article_question_section.dart';
+import 'package:flutter_ecoapp/views/widgets/articles/articleview/article_store_description_section.dart';
 import 'package:flutter_ecoapp/views/widgets/articles/full_eco_indicator.dart';
 import 'package:flutter_ecoapp/views/widgets/bottom_nav_bar.dart';
 import 'package:flutter_ecoapp/views/widgets/normal_button.dart';
